infra/requirements.txt
aws-cdk.core
aws-cdk.aws_s3
aws-cdk.aws_certificatemanager
aws-cdk.aws_cloudfront
aws-cdk.aws_route53
aws-cdk.aws_route53_targets
